backend/requirements.txt
# FastAPI and Web Framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0

# Data Processing and Validation
pydantic>=2.5.0
python-multipart>=0.0.6

# Database drivers
psycopg2-binary>=2.9.7
asyncpg>=0.29.0

# File type detection (optional - will fallback to mimetypes if not available)
# python-magic>=0.4.27
# python-magic-bin>=0.4.14; sys_platform == "win32"

# Environment and Configuration
python-dotenv>=1.0.0

# Date and Time
python-dateutil>=2.8.2

# HTTP Client for API integrations
httpx>=0.25.0
aiofiles>=23.2.1

# AI and ML Libraries
openai>=1.3.0
# llamaindex>=0.9.0  # Uncomment when implementing LlamaIndex integration

# Qdrant Vector Database
qdrant-client>=1.7.0

# Sentence Transformers for local embeddings
sentence-transformers>=2.2.2

# LangChain integration for RAG
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-qdrant>=0.1.0
langchain-community>=0.0.20

# Azure Services
azure-ai-documentintelligence>=1.0.0b1
azure-storage-blob>=12.19.0
azure-identity>=1.15.0

# Scientific Computing
numpy>=1.24.0
scipy>=1.11.0

# File Processing
PyPDF2>=3.0.1
python-docx>=1.1.0
openpyxl>=3.1.2

# Database - SQLite and PostgreSQL Implementation
sqlalchemy>=2.0.0
asyncpg>=0.29.0
databases[postgresql]>=0.8.0
aiosqlite>=0.19.0
alembic>=1.12.0

# Testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0  # For testing FastAPI endpoints

# Development tools
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
